Fixing build failure due to cross-module test deps
