+++ eureka-core/src/test/java/com/netflix/eureka/resources/ResponseCacheTest.java
@@ -1,7 +1,7 @@
 package com.netflix.eureka.resources;
 
 import com.netflix.blitz4j.LoggingConfiguration;
-import com.netflix.discovery.AbstractDiscoveryClientTester;
+import com.netflix.eureka.AbstractTester;
 import com.netflix.eureka.DefaultEurekaServerConfig;
 import com.netflix.eureka.EurekaServerConfig;
 import com.netflix.eureka.EurekaServerConfigurationManager;
@@ -14,7 +14,9 @@ import org.junit.Test;
 /**
  * @author Nitesh Kant
  */
-public class ResponseCacheTest extends AbstractDiscoveryClientTester {
+public class ResponseCacheTest extends AbstractTester {
+
+    private static final String REMOTE_REGION = "myremote";
 
     @Override
     @Before
@@ -30,12 +32,12 @@ public class ResponseCacheTest extends AbstractDiscoveryClientTester {
         EurekaServerConfig serverConfig = new DefaultEurekaServerConfig();
         EurekaServerConfigurationManager.getInstance().setConfiguration(serverConfig);
         ResponseCache cache = ResponseCache.getInstance();
-        ResponseCache.Key key = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, LOCAL_REGION_APP_NAME,
+        ResponseCache.Key key = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, REMOTE_REGION_APP_NAME,
                                                       ResponseCache.KeyType.JSON, Version.V1);
         String response = cache.get(key);
         Assert.assertNotNull("Cache get returned null.", response);
 
-        PeerAwareInstanceRegistry.getInstance().cancel(LOCAL_REGION_APP_NAME, LOCAL_REGION_INSTANCE_1_HOSTNAME, true);
+        PeerAwareInstanceRegistry.getInstance().cancel(REMOTE_REGION_APP_NAME, REMOTE_REGION_INSTANCE_1_HOSTNAME, true);
 
         Assert.assertNull("Cache after invalidate did not return null.", cache.get(key));
     }
@@ -45,13 +47,13 @@ public class ResponseCacheTest extends AbstractDiscoveryClientTester {
         EurekaServerConfig serverConfig = new DefaultEurekaServerConfig();
         EurekaServerConfigurationManager.getInstance().setConfiguration(serverConfig);
         ResponseCache cache = ResponseCache.getInstance();
-        ResponseCache.Key key = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, LOCAL_REGION_APP_NAME,
+        ResponseCache.Key key = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, REMOTE_REGION_APP_NAME,
                                                       new String[] {REMOTE_REGION},
                                                       ResponseCache.KeyType.JSON, Version.V1);
 
         Assert.assertNotNull("Cache get returned null.", cache.get(key));
 
-        PeerAwareInstanceRegistry.getInstance().cancel(LOCAL_REGION_APP_NAME, LOCAL_REGION_INSTANCE_1_HOSTNAME, true);
+        PeerAwareInstanceRegistry.getInstance().cancel(REMOTE_REGION_APP_NAME, REMOTE_REGION_INSTANCE_1_HOSTNAME, true);
 
         Assert.assertNull("Cache after invalidate did not return null.", cache.get(key));
     }
@@ -61,17 +63,17 @@ public class ResponseCacheTest extends AbstractDiscoveryClientTester {
         EurekaServerConfig serverConfig = new DefaultEurekaServerConfig();
         EurekaServerConfigurationManager.getInstance().setConfiguration(serverConfig);
         ResponseCache cache = ResponseCache.getInstance();
-        ResponseCache.Key key1 = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, LOCAL_REGION_APP_NAME,
+        ResponseCache.Key key1 = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, REMOTE_REGION_APP_NAME,
                                                       new String[] {REMOTE_REGION, "myregion2"},
                                                       ResponseCache.KeyType.JSON, Version.V1);
-        ResponseCache.Key key2 = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, LOCAL_REGION_APP_NAME,
+        ResponseCache.Key key2 = new ResponseCache.Key(ResponseCache.Key.EntityType.Application, REMOTE_REGION_APP_NAME,
                                                       new String[] {REMOTE_REGION},
                                                       ResponseCache.KeyType.JSON, Version.V1);
 
         Assert.assertNotNull("Cache get returned null.", cache.get(key1));
         Assert.assertNotNull("Cache get returned null.", cache.get(key2));
 
-        PeerAwareInstanceRegistry.getInstance().cancel(LOCAL_REGION_APP_NAME, LOCAL_REGION_INSTANCE_1_HOSTNAME, true);
+        PeerAwareInstanceRegistry.getInstance().cancel(REMOTE_REGION_APP_NAME, REMOTE_REGION_INSTANCE_1_HOSTNAME, true);
 
         Assert.assertNull("Cache after invalidate did not return null.", cache.get(key1));
         Assert.assertNull("Cache after invalidate did not return null.", cache.get(key2));
